% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{DM}
\alias{DM}
\title{Distance Matrix}
\usage{
DM(input_file, sequence_length, output_file)
}
\arguments{
\item{input_file}{a fasta file containing aligned amino acid sequences}

\item{sequence_length}{length of sequences in fasta file}

\item{output_file}{the name for the output .csv}
}
\value{
a .csv file containing a distance matrix
}
\description{
Calculates the Antigenic Hamming Distances between our sequences and stores the values as a distance matrix.

Required packages: stringdist, seqinr
}
\examples{
setwd("~/ssTA/tutorial/")
myDM <- DM("Output_GOOD/PracticeSeqs_Aligned_GOOD.fasta", 575, "DM_GOOD/PracticeSeqs_GOOD")

}
