% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_DR.R
\name{DM_DR}
\alias{DM_DR}
\title{Distance Matrix Dimension Reduction}
\usage{
DM_DR(input_file, output_file)
}
\arguments{
\item{input_file}{a distance matrix}

\item{output_file}{the name and/or directory path for the output .pdf and .csv}
}
\value{
a .csv file containing a distance matrix and a .pdf file which contains the plotted points
}
\description{
Performs multiple dimensional scaling with CMDscale, writes a .csv file containing coordinates, and produces a plot that represents the antigenic distances between sequences.

Required package: stringdist
}
\examples{
setwd("~/ssTA/tutorial/")
myDF <- DM_DR("DM_GOOD/PracticeSeqs_GOOD_DM.csv", "Output_GOOD/PracticeSeqs_points_GOOD")

}
