% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QC}
\alias{QC}
\title{Quality Control}
\usage{
QC(fasta_file, seq_length, output_name)
}
\arguments{
\item{fasta_file}{a fasta file containing nucleic acid sequences}

\item{seq_length}{length of sequences}

\item{output_name}{name for output fasta file}
}
\value{
a fasta file containing cleaned sequences is written
}
\description{
Performs Quality Control measures to nucleic acid sequence data. Removes any sequence which has a letter other than "ATCG".
}
\examples{
setwd("~/ssTA/tutorial/")
QC("Seqs.fasta", 1725, "CleanedSeqs.fasta")

}
