% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TA.R
\name{TA}
\alias{TA}
\title{Trait Association}
\usage{
TA(trait_col, path, Plot_name, Stats_name)
}
\arguments{
\item{trait_col}{column name which contains a trait to be analyzed}

\item{path}{directory path for output files}

\item{Plot_name}{name for output plots}

\item{Stats_name}{name for output stats file}
}
\value{
A .csv stats file containing anosim, adonis2, bDis.perm, and bDis.anova pvalues. A pdf file containing an Anosim plot. A pdf file containing a bDisper sdEllipse.
}
\description{
Performs 4 statistical tests for association between a trait and a distance matrix.
}
\examples{
setwd("~/ssTA/tutorial/DM/") ##set the directory path for the directory which contains all of your Distance Matrix files
myPheno<-read.csv("~/ssTA/tutorial/myPheno.csv") ##Read in metadata file
myPheno$name<- as.character(myPheno$name) ##should be character vector, sets GenBank.Accession col as character
TA(myPheno$subtype, "~/ssTA/tutorial/", "Test", "Test")
}
