% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Translate.R
\name{Translate}
\alias{Translate}
\title{Translate}
\usage{
Translate(input_file, output_file)
}
\arguments{
\item{input_file}{name of a fasta file containing Nucleic Acid sequences}

\item{output_file}{name of the output fasta file containing Amino Acid sequences}
}
\value{
a fasta file containing Amino acid sequences
}
\description{
Translates Nucleic acid sequences using bioseq seq_translate

Required package: bioseq
}
\examples{
setwd("~/ssTA/tutorial/")
myTranslatedSeqs <- Translate("Output_GOOD/PracticeSeqs_Cleaned_GOOD.fasta", "Output_GOOD/PracticeSeqs_Translated_GOOD.fasta")

}
