% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QC}
\alias{QC}
\title{Quality Control}
\usage{
QC(fasta_file, seq_length_vector, output_name)
}
\arguments{
\item{fasta_file}{a fasta file containing nucleic acid sequences}

\item{seq_length_vector}{a vector containing the lengths of nucleic acid sequences in the fasta file}

\item{output_name}{name for output fasta file}
}
\value{
a list containing cleaned sequences is stored in the RStudio environment &  a fasta file containing cleaned sequences is written to the working directory.
}
\description{
Performs Quality Control measures to nucleic acid sequence data.

Pulls sequences of given lengths. Removes any sequence which has a character or symbol other than "ATCG".

Required package: seqinr
}
\examples{
setwd("~/ssTA/tutorial/")
myLengths <- c(1725)
myQCSeqs <- QC("PracticeSeqs_NA_GOOD.fasta", myLengths, "Output_GOOD/PracticeSeqs_Cleaned_GOOD.fasta")
myLengths <- c(1725, 1726, 1727, 1729)
myQCSeqs <- QC("PracticeSeqs_NA_MultiLengths.fasta", myLengths, "Output_MultiLengths/PracticeSeqs_Cleaned_MultiLengths.fasta")

}
