% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TA.R
\name{TA}
\alias{TA}
\title{Trait Association}
\usage{
TA(name_column_number, trait_column_number, output_path, Stats_name)
}
\arguments{
\item{name_column_number}{column number which contains names of sequences}

\item{trait_column_number}{column number which contains a trait to be analyzed}

\item{output_path}{directory path for output files}

\item{Stats_name}{name for output stats file}
}
\value{
A .csv stats file containing anosim, adonis2, bDis.perm, and bDis.anova pvalues. A pdf file containing an Anosim plot. A pdf file containing a bDisper sdEllipse.
}
\description{
Performs 4 statistical tests (anosim, adonis2, bDis.perm, and bDis.anova) for association between a trait and a distance matrix.

Required packages: Rtsne, vegan
}
\examples{
setwd("~/ssTA/tutorial/")
myPheno<-read.csv("myPheno.csv")
setwd("DM_GOOD")
myStats <- TA(1, 2, "../", "PracticeSeqs_GOOD")
setwd("..")
}
