% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TA.R
\name{TA}
\alias{TA}
\title{Trait Association}
\usage{
TA(trait_col_num, output_path, Plot_name, Stats_name)
}
\arguments{
\item{output_path}{directory path for output files}

\item{Plot_name}{name for output plots}

\item{Stats_name}{name for output stats file}

\item{trait_col}{column name which contains a trait to be analyzed}
}
\value{
A .csv stats file containing anosim, adonis2, bDis.perm, and bDis.anova pvalues. A pdf file containing an Anosim plot. A pdf file containing a bDisper sdEllipse.
}
\description{
Performs 4 statistical tests (anosim, adonis2, bDis.perm, and bDis.anova) for association between a trait and a distance matrix.

Required packages: Rtsne, vegan
}
\examples{
setwd("~/ssTA/tutorial/")
myPheno<-read.csv("myPheno.csv")
setwd("DM_GOOD")
myStats <- TA(2, "../Output_GOOD/", "PracticeSeqs_GOOD", "PracticeSeqs_GOOD")
setwd("..")
}
